core,w32: Silence compiler warnings.
* src/w32-util.c (_gpgme_w32_cancel_synchronous_io): Add some casts.
--

+++ src/w32-util.c
@@ -308,10 +308,11 @@ _gpgme_w32_cancel_synchronous_io (HANDLE thread)
 
   if (func)
     {
-      if (!func (thread) && GetLastError() != ERROR_NOT_FOUND)
+      if (!func ((DWORD)thread) && GetLastError() != ERROR_NOT_FOUND)
         {
           TRACE (DEBUG_ENGINE, "gpgme:CancelSynchronousIo", NULL,
-                  "called for thread %p: ec=%d", thread, GetLastError ());
+                 "called for thread %p: ec=%u",
+                 thread, (unsigned int)GetLastError ());
         }
     }
   else
